+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
+## [Unreleased] - 2021-12-03
+
+### Added
+- Automated deploy to AWS instance
+
 ## [1.2.10] - 2021-12-03
 
 ### Added